pyproject.toml
[project]
name = "solio-autosolve"
version = "0.1.0"
description = "Automated FPL optimization using Solio Analytics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.57.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
solio = "solio_autosolve.main:main"
solio-login = "solio_autosolve.login:main"
solio-solve = "solio_autosolve.solve:main"
solio-parse = "solio_autosolve.parser:main"
solio-email = "solio_autosolve.email_sender:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/solio_autosolve"]
